+++ CHANGELOG.md
@@ -6,6 +6,7 @@
   * `Mockery.Assertions.refute_called/3`
   * `Mockery.Assertions.assert_called/4`
   * `Mockery.Assertions.refute_called/4`
+* Removed unnecessary `Elixir.` namespace from module names in error messages
 
 ## 1.1.0
 * Added `Mockery.mock/2` (`Mockery.mock/3` value defaults to `:mocked`)